Advanced feature to control generation of tape filename
The deefault behaviour continues to name tape files as the name of the drive to mount, passed by the remote client.
The new options -R and -W allow to specify a printf-style format string (separately for READ-only  and WRITE or BOTh (R/W)
mounts where the client'S chaosnet peer address and the drive name can be used as filename components. Examples
are given in the usage help output.

The main use case could be a remote server that will prevent tapes from different client hosts to accidently overwrite
each other, but allowing remote hosts to exchange tapes in read-only mounts.